Rename run report utils
